requirements.txt
Pillow>=8.3.2
edq-utils>=0.0.5
